requirements.txt
cython==3.1.0
attr==0.3.2
colorama==0.4.6
ConfigParser==7.2.0
cryptography==45.0.1
docutils==0.21.2
filelock==3.18.0
HTMLParser==0.0.2
ipython==8.32.0
ipywidgets==8.1.7
keyring==25.6.0
openai==1.79.0
Pillow==11.2.1
protobuf==6.31.0
pyOpenSSL==25.1.0
python-dotenv==1.1.0
redis==6.1.0
reportlab==4.4.1
Sphinx==8.2.3
streamlit==1.45.1
streamlit_extras==0.7.1
thread==2.0.5
urllib3_secure_extra==0.1.0
reportlab==4.4.1 
xmlrpclib==1.0.1
jnius==1.1.0
